Add files via upload
This overall system works on WiFi connectivity and the Internet & it's connected with our VeggieFridge server and VeggieFridge Android App 

This alert will be shown on Android app notifications and also can send notifications through SMS. In this customers can also change the password from the Android app. The customer can also able to open the door from the Android app & we can add the extra feature which is we can check the milk quantity inside the bottle from the app.

It will be password protected for an authentication system based on the keyboard as well as the Android app and the status of the device will shown on LCD display.